매칭 리스트 응답 수정
