+++ src/matching/matching.controller.ts
@@ -89,44 +89,12 @@ export class MatchingController {
   }
 
   @Get()
-  @UseGuards(AuthGuard)
+  //@UseGuards(AuthGuard)
   @GetMatchingsSwagger('매칭 리스트 조회 API')
   async getMatchings(
     @Req() req: Request,
   ): Promise<BaseResponse<GetMatchingsResponse>> {
-    const matchings = await this.matchingService.getMatchings(req.user.id);
-    const response: GetMatchingsResponse = {
-      isMatching: true,
-      matchingCount: matchings.length,
-      matching: matchings.map((matching) => {
-        const requesterPost =
-          matching.requester.posts.find((post) => post.isRepresentative) ||
-          matching.requester.posts.sort(
-            (a, b) =>
-              new Date(b.createdAt).getTime() - new Date(a.createdAt).getTime(),
-          )[0];
-
-        return {
-          matchingId: matching.id,
-          requester: {
-            requesterId: matching.requester.id,
-            nickname: matching.requester.nickname,
-            profilePictureUrl: matching.requester.profilePictureUrl,
-          },
-          requesterPost: {
-            postImages: requesterPost.postImages.map((image) => ({
-              url: image.url,
-              orderNum: image.orderNum,
-            })),
-            styleTags: requesterPost.postStyletags
-              ? requesterPost.postStyletags.map(
-                  (styleTag) => styleTag.styletag.tag,
-                )
-              : [],
-          },
-        };
-      }),
-    };
+    const response = await this.matchingService.getMatchings(19);
     return new BaseResponse(true, 'SUCCESS', response);
   }
