+++ src/matching/matching.service.ts
@@ -11,6 +11,7 @@ import { ChatMessageService } from 'src/chat-message/chat-message.service';
 import { ChatRoom } from 'src/common/entities/chat-room.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { PatchMatchingRequestDto } from './dto/Patch-matching.request';
+import { GetMatchingsResponse } from './dto/get-matching.response';
 
 @Injectable()
 export class MatchingService {
@@ -85,20 +86,54 @@ export class MatchingService {
     }
   }
 
-  async getMatchings(currentUserId: number): Promise<Matching[]> {
-    return await this.matchingRepository.find({
-      where: {
-        requestStatus: 'pending',
-        status: 'activated',
-        target: { id: currentUserId },
-      },
-      relations: [
-        'requester',
-        'requester.posts',
-        'requester.posts.postImages',
-        'requester.posts.postStyletags.styletag',
-      ],
-    });
+  async getMatchings(currentUserId: number): Promise<GetMatchingsResponse> {
+    const matchings = await this.matchingRepository
+      .createQueryBuilder('matching')
+      .leftJoinAndSelect('matching.requester', 'requester')
+      .leftJoinAndSelect('requester.posts', 'post')
+      .leftJoinAndSelect('post.postImages', 'image')
+      .leftJoinAndSelect('post.postStyletags', 'styleTag')
+      .leftJoinAndSelect('styleTag.styletag', 'styletag')
+      .where('matching.targetId = :currentUserId', { currentUserId })
+      .andWhere('matching.requestStatus = :status', { status: 'pending' })
+      .andWhere('matching.status = :activated', { activated: 'activated' })
+      .orderBy(
+        // 우선순위: isRepresentative가 true인 게시물 먼저, 그 다음은 최신 게시물
+        'CASE WHEN post.isRepresentative = true THEN 0 ELSE 1 END',
+        'ASC',
+      )
+      .addOrderBy('post.createdAt', 'DESC')
+      .getMany();
+
+    const response: GetMatchingsResponse = {
+      isMatching: true,
+      matchingCount: matchings.length,
+      matching: matchings.map((matching) => {
+        const requesterPost = matching.requester.posts[0]; // 첫 번째 게시물이 조건에 맞는 게시물
+
+        return {
+          matchingId: matching.id,
+          requester: {
+            requesterId: matching.requester.id,
+            nickname: matching.requester.nickname,
+            profilePictureUrl: matching.requester.profilePictureUrl,
+          },
+          requesterPost: {
+            postImages: requesterPost.postImages.map((image) => ({
+              url: image.url,
+              orderNum: image.orderNum,
+            })),
+            styleTags: requesterPost.postStyletags
+              ? requesterPost.postStyletags.map(
+                  (styleTag) => styleTag.styletag.tag,
+                )
+              : [],
+          },
+        };
+      }),
+    };
+
+    return response;
   }
 
   async getMatchingById(matchingId: number): Promise<Matching> {
